+++ src/renderer/areaChart.js
@@ -7,7 +7,7 @@
 
 import { buildEnvelope } from './_envelope.js';
 import { validateValues, validateColorSpec } from '../helpers/validate.js';
-import { gradient, fadeToTransparent, fill } from '../helpers/colors.js';
+import { gradient, fadeToTransparent, fill, APP_DEFAULT_BACKGROUND, APP_DEFAULT_BACKGROUND_TYPE } from '../helpers/colors.js';
 
 /**
  * @param {object} opts
@@ -76,8 +76,8 @@ export function makeAreaChart({
   yMax = null,
   nameFont = null,
   namePosition = 'TOP',
-  backgroundColor = null,
-  backgroundType = 'NONE',
+  backgroundColor = APP_DEFAULT_BACKGROUND,
+  backgroundType = APP_DEFAULT_BACKGROUND_TYPE,
   borderColor = null,
   shadowColor = null,
   refreshInterval = 30,