+++ src/renderer/circularProgress.js
@@ -8,7 +8,7 @@
 
 import { buildEnvelope } from './_envelope.js';
 import { validateProgress, validateColorSpec } from '../helpers/validate.js';
-import { gradient, fill } from '../helpers/colors.js';
+import { gradient, fill, APP_DEFAULT_BACKGROUND, APP_DEFAULT_BACKGROUND_TYPE } from '../helpers/colors.js';
 
 /**
  * @param {object} opts
@@ -48,8 +48,8 @@ export function makeCircularProgress({
   showGlow = false,
   nameFont = null,
   namePosition = 'TOP',
-  backgroundColor = null,
-  backgroundType = 'NONE',
+  backgroundColor = APP_DEFAULT_BACKGROUND,
+  backgroundType = APP_DEFAULT_BACKGROUND_TYPE,
   borderColor = null,
   shadowColor = null,
   refreshInterval = 30,