+++ src/renderer/kpiMetric.js
@@ -7,7 +7,7 @@
 
 import { buildEnvelope } from './_envelope.js';
 import { validateColorSpec } from '../helpers/validate.js';
-import { gradient, fill, fadeToTransparent } from '../helpers/colors.js';
+import { gradient, fill, fadeToTransparent, APP_DEFAULT_BACKGROUND, APP_DEFAULT_BACKGROUND_TYPE } from '../helpers/colors.js';
 
 /**
  * @param {object} opts
@@ -49,8 +49,8 @@ export function makeKPIMetric({
   showGlow = false,
   nameFont = null,
   namePosition = 'TOP',
-  backgroundColor = null,
-  backgroundType = 'NONE',
+  backgroundColor = APP_DEFAULT_BACKGROUND,
+  backgroundType = APP_DEFAULT_BACKGROUND_TYPE,
   borderColor = null,
   shadowColor = null,
   refreshInterval = 30,