+++ src/renderer/pieChart.js
@@ -6,7 +6,7 @@
  */
 
 import { buildEnvelope } from './_envelope.js';
-import { gradient } from '../helpers/colors.js';
+import { gradient, APP_DEFAULT_BACKGROUND, APP_DEFAULT_BACKGROUND_TYPE } from '../helpers/colors.js';
 
 /**
  * @param {object} opts
@@ -38,8 +38,8 @@ export function makePieChart({
   showPercentage = false,
   nameFont = null,
   namePosition = 'TOP',
-  backgroundColor = null,
-  backgroundType = 'NONE',
+  backgroundColor = APP_DEFAULT_BACKGROUND,
+  backgroundType = APP_DEFAULT_BACKGROUND_TYPE,
   borderColor = null,
   shadowColor = null,
   refreshInterval = 30,